SaterBot-main/requirements.txt
discord.py[voice]==1.5
discord.py==1.5
youtube-dl==2020.7.28
websockets==8.1
aiohttp==3.6.2
PyNaCl==1.3.0
multidict==4.7.5
async-timeout==3.0.1
yarl==1.4.2
attrs==19.3.0
chardet==3.0.4
cffi==1.14.0
six==1.15.0
idna==2.10
pycparser==2.20
pymongo==3.10.1
dnspython==1.16.0
aiohttp==3.6.2
wheel==0.34.2
google-api-python-client==1.9.3
asyncio
psutil
pytz
timeago==1.0.13
requests
colorthief
safygiphy
emoji
discord_webhook
